+++ src/NuGet.Clients/NuGet.CommandLine/Commands/DeleteCommand.cs
@@ -46,7 +46,6 @@ namespace NuGet.CommandLine
             //TODO: Consider a better resource name, like PackageUpdaterResource
             //Do it after the common hander resource is available, to avoid throw away code.
             PushCommandResource pushCommandResource = await sourceRepository.GetResourceAsync<PushCommandResource>();
-            var packageUpdater = pushCommandResource.GetPackageUpdater();
 
             //If the user did not pass an API Key look in the config file
             string apiKey = GetApiKey(source);
@@ -59,9 +58,8 @@ namespace NuGet.CommandLine
             if (NonInteractive || Console.Confirm(String.Format(CultureInfo.CurrentCulture, LocalizedResourceManager.GetString("DeleteCommandConfirm"), packageId, packageVersion, sourceDisplayName)))
             {
                 Console.WriteLine(LocalizedResourceManager.GetString("DeleteCommandDeletingPackage"), packageId, packageVersion, sourceDisplayName);
-                var userAgent = UserAgent.CreateUserAgentString(CommandLineConstants.UserAgent);
                 //TODO: confirm, no timeout on delete command?
-                await packageUpdater.DeletePackage(apiKey, packageId, packageVersion, userAgent, Console, CancellationToken.None);
+                await pushCommandResource.DeletePackage(apiKey, packageId, packageVersion, Console, CancellationToken.None);
                 Console.WriteLine(LocalizedResourceManager.GetString("DeleteCommandDeletedPackage"), packageId, packageVersion);
             }
             else
