+++ src/NuGet.Core/NuGet.Protocol.Core.v2/FactoryExtensionsV2.cs
@@ -26,7 +26,6 @@ namespace NuGet.Protocol.Core.v2
             yield return new Lazy<INuGetResourceProvider>(() => new MetadataResourceV2Provider());
             yield return new Lazy<INuGetResourceProvider>(() => new PackageRepositoryResourceV2Provider());
             yield return new Lazy<INuGetResourceProvider>(() => new ListCommandResourceV2Provider());
-            yield return new Lazy<INuGetResourceProvider>(() => new PushCommandResourceV2Provider());
             yield break;
         }
     }
