+++ src/NuGet.Core/NuGet.Protocol.Core.v3/FactoryExtensionsV3.cs
@@ -39,6 +39,7 @@ namespace NuGet.Protocol.Core.v3
             yield return new Lazy<INuGetResourceProvider>(() => new LocalV3FindPackageByIdResourceProvider());
             yield return new Lazy<INuGetResourceProvider>(() => new LocalV2FindPackageByIdResourceProvider());
             yield return new Lazy<INuGetResourceProvider>(() => new ListCommandResourceV3Provider());
+            yield return new Lazy<INuGetResourceProvider>(() => new PushCommandResourceV2Provider());
             yield return new Lazy<INuGetResourceProvider>(() => new PushCommandResourceV3Provider());
             yield break;
         }
