+++ src/NuGet.Core/NuGet.Protocol.Core.v3/PushCommandResourceV3Provider.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Linq;
-using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using NuGet.Protocol.Core.Types;
@@ -31,30 +30,20 @@ namespace NuGet.Protocol.Core.v3
                 // Returning null here will result in ListCommandResource
                 // getting returned for this very v3 package source as if it was a v2 package source
                 var baseUrl = serviceIndex[ServiceTypes.PackagePublish].FirstOrDefault();
-                pushCommandResource = new PushCommandResource(baseUrl?.AbsoluteUri, 
-                    async (baseUri, cancellationToken) => {
-                        ICredentials credentials = null;
-                        if (HttpHandlerResourceV3.PromptForCredentials != null)
-                        {
-                            credentials = await HttpHandlerResourceV3.PromptForCredentials(baseUri, cancellationToken);
-                        }
-                        return credentials;
-                    },
-                    (baseUri, credentials) => {
-                        if (HttpHandlerResourceV3.CredentialsSuccessfullyUsed != null && credentials != null)
-                        { 
-                            HttpHandlerResourceV3.CredentialsSuccessfullyUsed(baseUri, credentials);
-                        }
 
-                    },
-                    async () => (await source.GetResourceAsync<HttpHandlerResource>(token))
-                 );
+                HttpSource httpSource = null;
+                string sourceUri = baseUrl?.AbsoluteUri;
+                if (!string.IsNullOrEmpty(sourceUri) && !(new Uri(sourceUri)).IsFile)
+                {
+                    var httpSourceResource = await source.GetResourceAsync<HttpSourceResource>(token);
+                    httpSource = httpSourceResource.HttpSource;
+                }
+
+                pushCommandResource = new PushCommandResource(sourceUri, httpSource);
             }
 
             var result = new Tuple<bool, INuGetResource>(pushCommandResource != null, pushCommandResource);
             return result;
         }
-
-        
     }
 }
